Expose URLs to static CloudFiles storage to templates
This context processor allows drop-in replacement of local static
files with a CloudFiles CDN storage. The SSL URL is also provided
as a context variable in case it is explicitly required. In case
some files are stored locally in the STATIC_URL, it can be accessed
through the LOCAL_STATIC_URL context variable.